Support MacOS compilation
- Added platform-specific macro definitions
- Use char and wchar_t if unicode character is not supported
  - XCode library doesn't support unicode character correctly (locale)
- Removed inline in platform definitions

+++ Source/Base/Private/CubeFormat.cpp
@@ -30,7 +30,7 @@ namespace cube
                 return (offset + (alignment - 1)) & ~(alignment - 1);
             }
 
-            void* Allocate(size_t n)
+            void* Allocate(Uint64 n)
             {
                 Uint64 allocatedSize = (Uint64)thlFormatMemoryBlock.currentPtr - (Uint64)thlFormatMemoryBlock.startPtr;
                 if (allocatedSize + n > blockSize)
@@ -45,7 +45,7 @@ namespace cube
                 return ptr;
             }
 
-            void* Allocate(size_t n, size_t alignment)
+            void* Allocate(Uint64 n, Uint64 alignment)
             {
                 Uint64 currentOffset = (Uint64)thlFormatMemoryBlock.currentPtr;
